# docs/swagger.yaml
basePath: /
definitions:
  models.PropertySwagger:
    properties:
      description:
        example: Spacious apartment near downtown.
        type: string
      liked_by:
        items:
          type: string
        type: array
      location:
        example: New York
        type: string
      owner_email:
        example: owner@example.com
        type: string
      owner_name:
        example: John Doe
        type: string
      owner_pic:
        example: https://example.com/pic.jpg
        type: string
      pictures:
        items:
          type: string
        type: array
      price:
        example: 2500
        type: number
      thumbnail:
        type: string
      title:
        example: Modern 2BHK Apartment
        type: string
    type: object
  models.UserSwagger:
    properties:
      email:
        example: user@example.com
        type: string
      liked_properties:
        items:
          type: string
        type: array
      location:
        type: string
      name:
        example: Alice Smith
        type: string
      posted_properties:
        items:
          type: string
        type: array
      preferred_location:
        type: string
      profile_pic:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server.
  title: Dwello-api
  version: "1.0"
paths:
...
  /api/users/{email}:
    get:
      description: Get a user document based on email address
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserSwagger'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User by Email
      tags:
      - Users
  /api/users/{email}/liked-properties:
    get:
      description: Get a list of properties the user has liked
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PropertySwagger'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Liked Properties
      tags:
      - Users
  /api/users/{email}/location:
    put:
      consumes:
      - application/json
      description: Update the location field of a user
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      - description: Location JSON
        in: body
        name: location
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update User Location
      tags:
      - Users
  /api/users/{email}/posted-properties:
    get:
      description: Get a list of properties the user has posted
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PropertySwagger'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Posted Properties
      tags:
      - Users
...
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user or return existing user if already registered
      parameters:
      - description: User JSON
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserSwagger'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserSwagger'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register or Login User
      tags:
      - Users
swagger: "2.0"